rename: 컴포넌트 디렉토리 위치 변경
